+++ examples/lists.js
@@ -280,7 +280,7 @@ var docDefinition = {
 			ul: [
 				'item 1',
 				'item 2',
-				'item 3'
+				{ text: 'item 3 with custom marker color', markerColor: 'lime' }
 			]
 		},
 		{ text: '\n\nColored ordered list', style: 'header' },
@@ -299,7 +299,7 @@ var docDefinition = {
 			ol: [
 				'item 1',
 				'item 2',
-				'item 3'
+				{ text: 'item 3 with custom marker color', markerColor: 'lime' }
 			]
 		},
 		{ text: '\n\nOrdered list - type: lower-alpha', style: 'header' },
@@ -393,6 +393,23 @@ var docDefinition = {
 				'item 1',
 				{ text: 'item 2', listType: 'none' },
 				{ text: 'item 3', listType: 'circle' }
+			]
+		},
+		{ text: '\n\nOrdered list with an empty item:', style: 'header' },
+		{
+			ol: [
+				'First item',
+				'Second item',
+				'',
+				'Fourth item'
+			]
+		},
+		{ text: '\n\nUnordered list with an empty item:', style: 'header' },
+		{
+			ul: [
+				'First bullet',
+				'',
+				'Third bullet'
 			]
 		},
 	],