+++ src/printer.js
@@ -15,6 +15,7 @@ var isNumber = require('./helpers').isNumber;
 var isBoolean = require('./helpers').isBoolean;
 var isArray = require('./helpers').isArray;
 var isUndefined = require('./helpers').isUndefined;
+var isObject = require('./helpers').isObject;
 var isPattern = require('./helpers').isPattern;
 var getPattern = require('./helpers').getPattern;
 var SVGtoPDF = require('./3rd-party/svg-to-pdfkit');
@@ -110,6 +111,14 @@ function PdfPrinter(fontDescriptors) {
  */
 PdfPrinter.prototype.createPdfKitDocument = function (docDefinition, options) {
 	options = options || {};
+
+	if (!isObject(docDefinition)) {
+		throw new Error("Parameter 'docDefinition' has an invalid type. Object expected.");
+	}
+
+	if (!isObject(options)) {
+		throw new Error("Parameter 'options' has an invalid type. Object expected.");
+	}
 
 	docDefinition.version = docDefinition.version || '1.3';
 	docDefinition.subset = docDefinition.subset || undefined;
@@ -769,7 +778,7 @@ function renderImage(image, x, y, pdfKitDoc) {
 }
 
 function renderSVG(svg, x, y, pdfKitDoc, fontProvider) {
-	var options = Object.assign({ width: svg._width, height: svg._height, assumePt: true }, svg.options);
+	var options = Object.assign({ width: svg._width, height: svg._height, assumePt: true, useCSS: !isString(svg.svg) }, svg.options);
 	options.fontCallback = function (family, bold, italic) {
 		var fontsFamily = family.split(',').map(function (f) { return f.trim().replace(/('|")/g, ''); });
 		var font = findFont(fontProvider.fonts, fontsFamily, svg.font || 'Roboto');