+++ tests/textTools.js
@@ -139,6 +139,7 @@ describe('TextTools', function () {
 			font: 'Helvetica'
 		});
 	var styleStackNoWrap = new StyleContextStack({}, { noWrap: true });
+	var styleStackWordBreakBreakAll = new StyleContextStack({}, { wordBreak: 'break-all' });
 
 	describe('splitWords', function () {
 		it('should do basic splitting', function () {
@@ -241,13 +242,22 @@ describe('TextTools', function () {
 		it('should support cast to text', function () {
 			var node = docPreprocessor.preprocessNode(mixedTextArrayWithVariousTypes);
 			var result = TextTools.__get__('normalizeTextArray')(node.stack, styleStack);
-			assert.equal(result.length, 6);
-			assert.equal(result[0].text, '2016');
-			assert.equal(result[1].text, 'true');
-			assert.equal(result[2].text, 'false');
-			assert.equal(result[3].text, '2016');
-			assert.equal(result[4].text, 'true');
-			assert.equal(result[5].text, 'false');
+			assert.equal(result.length, 10);
+			assert.equal(result[0].text, '');
+			assert.equal(result[1].text, '');
+			assert.equal(result[2].text, '2016');
+			assert.equal(result[3].text, 'true');
+			assert.equal(result[4].text, 'false');
+			assert.equal(result[5].text, '');
+			assert.equal(result[6].text, '');
+			assert.equal(result[7].text, '2016');
+			assert.equal(result[8].text, 'true');
+			assert.equal(result[9].text, 'false');
+		});
+
+		it('should support wordBreak break-all', function () {
+			var result = TextTools.__get__('normalizeTextArray')(sampleText, styleStackWordBreakBreakAll);
+			assert.equal(result.length, 76);
 		});
 
 		it('should support keep noWrap from style', function () {